address PR comments: - remove I from non-interface class - remove engine getter (at first only temporarily to see, if we can live without it) - cleanup CMakeList.txt
